feat: improve handling of api response in flask app

+++ ichrisbirch/app/routes/autotasks.py
@@ -7,7 +7,7 @@ from fastapi import status
 from flask import Blueprint, Response, flash, render_template, request
 
 from ichrisbirch import schemas
-from ichrisbirch.app.helpers import log_flash_raise_error
+from ichrisbirch.app.helpers import handle_if_not_response_code
 from ichrisbirch.config import get_settings
 from ichrisbirch.models.autotask import TaskFrequency
 from ichrisbirch.models.task import TaskCategory
@@ -29,42 +29,36 @@ def index():
         data: dict[str, Any] = request.form.to_dict()
         method = data.pop('method')
         logger.debug(f'{request.referrer=} | {method=} | {data=}')
-        match method:
-            case 'add':
-                try:
-                    autotask = schemas.AutoTaskCreate(**data)
-                except pydantic.ValidationError as e:
-                    logger.exception(e)
-                    flash(str(e), 'error')
-                    return Response(str(e), status=status.HTTP_400_BAD_REQUEST)
-                response = requests.post(AUTOTASKS_API_URL, data=autotask.model_dump_json(), timeout=TIMEOUT)
-                if response.status_code != status.HTTP_201_CREATED:
-                    log_flash_raise_error(response, logger)
-                else:  # Run the new autotask
-                    task_response = requests.get(
-                        f'{AUTOTASKS_API_URL}/{response.json().get("id")}/run/', timeout=TIMEOUT
-                    )
-                    if task_response.status_code != status.HTTP_200_OK:
-                        log_flash_raise_error(task_response, logger)
 
-            case 'run':
-                id = request.form.get('id')
-                response = requests.get(f'{AUTOTASKS_API_URL}/{id}/run/', timeout=TIMEOUT)
-                if response.status_code != status.HTTP_200_OK:
-                    log_flash_raise_error(response, logger)
+        if method == 'add':
+            try:
+                autotask = schemas.AutoTaskCreate(**data)
+            except pydantic.ValidationError as e:
+                logger.exception(e)
+                flash(str(e), 'error')
+                return Response(str(e), status=status.HTTP_400_BAD_REQUEST)
+            response = requests.post(AUTOTASKS_API_URL, data=autotask.model_dump_json(), timeout=TIMEOUT)
+            handle_if_not_response_code(201, response, logger)
+            # Run the new autotask
+            new_autotask_run_url = f'{AUTOTASKS_API_URL}/{response.json().get("id")}/run/'
+            task_response = requests.get(new_autotask_run_url, timeout=TIMEOUT)
+            handle_if_not_response_code(200, task_response, logger)
 
-            case 'delete':
-                autotask_id = data.get('id')
-                response = requests.delete(f'{AUTOTASKS_API_URL}/{autotask_id}/', timeout=TIMEOUT)
-                if response.status_code != status.HTTP_204_NO_CONTENT:
-                    log_flash_raise_error(response, logger)
+        elif method == 'run':
+            id = request.form.get('id')
+            response = requests.get(f'{AUTOTASKS_API_URL}/{id}/run/', timeout=TIMEOUT)
+            handle_if_not_response_code(200, response, logger)
 
-            case _:
-                return Response(f'Method {method} not allowed', status=status.HTTP_405_METHOD_NOT_ALLOWED)
+        elif method == 'delete':
+            autotask_id = data.get('id')
+            response = requests.delete(f'{AUTOTASKS_API_URL}/{autotask_id}/', timeout=TIMEOUT)
+            handle_if_not_response_code(204, response, logger)
+
+        else:
+            return Response(f'Method {method} not allowed', status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
     response = requests.get(AUTOTASKS_API_URL, timeout=TIMEOUT)
-    if response.status_code != status.HTTP_200_OK:
-        log_flash_raise_error(response, logger)
+    handle_if_not_response_code(200, response, logger)
     autotasks = [schemas.AutoTask(**task) for task in response.json()]
     return render_template(
         'autotasks/index.html', autotasks=autotasks, task_categories=TASK_CATEGORIES, task_frequencies=TASK_FREQUENCIES
