+++ ichrisbirch/app/routes/countdowns.py
@@ -7,7 +7,7 @@ from fastapi import status
 from flask import Blueprint, Response, flash, render_template, request
 
 from ichrisbirch import schemas
-from ichrisbirch.app.helpers import log_flash_raise_error
+from ichrisbirch.app.helpers import handle_if_not_response_code
 from ichrisbirch.config import get_settings
 
 settings = get_settings()
@@ -24,29 +24,28 @@ def index():
     if request.method == 'POST':
         data: dict[str, Any] = request.form.to_dict()
         method = data.pop('method')
-        logger.debug(f'{request.referrer=} | {method=} | {data=}')
-        match method:
-            case 'add':
-                try:
-                    countdown = schemas.CountdownCreate(**data).model_dump_json()
-                except pydantic.ValidationError as e:
-                    logger.exception(e)
-                    flash(str(e), 'error')
-                    return Response(str(e), status=status.HTTP_400_BAD_REQUEST)
-                response = requests.post(COUNTDOWNS_API_URL, data=countdown, timeout=TIMEOUT)
-                if response.status_code != status.HTTP_201_CREATED:
-                    log_flash_raise_error(response, logger)
+        logger.debug(f'{request.referrer=} {method=}')
+        logger.debug(f'{data=}')
 
-            case 'delete':
-                id = data.get('id')
-                response = requests.delete(f'{COUNTDOWNS_API_URL}/{id}', timeout=TIMEOUT)
-                if response.status_code != status.HTTP_204_NO_CONTENT:
-                    log_flash_raise_error(response, logger)
-            case _:
-                return Response(f'Method {method} not allowed', status=status.HTTP_405_METHOD_NOT_ALLOWED)
+        if method == 'add':
+            try:
+                countdown = schemas.CountdownCreate(**data).model_dump_json()
+            except pydantic.ValidationError as e:
+                logger.exception(e)
+                flash(str(e), 'error')
+                return Response(str(e), status=status.HTTP_400_BAD_REQUEST)
+            response = requests.post(COUNTDOWNS_API_URL, data=countdown, timeout=TIMEOUT)
+            handle_if_not_response_code(201, response, logger)
+
+        elif method == 'delete':
+            id = data.get('id')
+            response = requests.delete(f'{COUNTDOWNS_API_URL}/{id}', timeout=TIMEOUT)
+            handle_if_not_response_code(204, response, logger)
+
+        else:
+            return Response(f'Method {method} not allowed', status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
     response = requests.get(COUNTDOWNS_API_URL, timeout=TIMEOUT)
-    if response.status_code != status.HTTP_200_OK:
-        log_flash_raise_error(response, logger)
+    handle_if_not_response_code(200, response, logger)
     countdowns = [schemas.Countdown(**countdown) for countdown in response.json()]
     return render_template('countdowns/index.html', countdowns=countdowns)
